+++ apps/web/src/features/auth/model/mutations/index.ts
@@ -1,6 +1,5 @@
 export * from './useSessionMutations';
 export * from './useRequestNativeSocialTokenMutation';
-export * from './useNativeSocialLoginMutation';
 export * from './useNativeSocialLoginFlowMutation';
 export * from './useSocialRegistrationMutation';
 export * from './useSendEmailVerificationCodeMutation';
